src/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-js-asset==2.1.0
django-mptt==0.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
install==1.3.5
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
text-unidecode==1.3
uritemplate==4.1.1
